git: name formatting change
For posterity: Jōshin is my Dharma name. I received it in 2018. April 1,
Easter Sunday, 2018, to be exact. It is 成真 in kanji. It means “fulfill
authenticity.”

For a while I was doing full-time Buddhist monastic training, and during
that time I was going by that name everywhere all the time. Now that I’m
a householder and doing other stuff, I’m more typically defaulting to my
legal name. But I like all of my names, and I’m happy for you to call me
any of them — including “mrdomino.”