nvim: remove lsp-zero.nvim
lsp-zero.nvim says it's going away since neovim aims to be resolving the
issues for which it existed. Drop it and try to put in the configuration
it advises instead. Also removes LuaSnip since vim.snippet apparently is
now available, and luvit-meta since it seems not to be needed anymore.